# docs/api/erp-v1-openapi.yaml
openapi: 3.1.0
info:
  title: ERP Operational Core API
  version: 1.0.0
  description: "Phase 1 production contract. Lifecycle and scope live in operational-core-v1-spec.md; shared rules live in api-conventions.md."
servers:
  - url: https://erp.example.invalid
    description: "Same-site reverse proxy that forwards /api to the backend."
paths:
...
  /api/v1/stock-positions:
    get:
      operationId: listStockPositions
      summary: "List current finished-goods Stock Positions"
      description: "Filters: customerId, productionOrderId, inStock. Sort allowlist is currentQty, updatedAt and id."
      security: [{ bearerAuth: [] }]
      x-required-permission: "STOCK:VIEW"
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - { name: customerId, in: query, schema: { type: string, format: uuid } }
        - { name: productionOrderId, in: query, schema: { type: string, format: uuid } }
        - { name: inStock, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: "Paged Stock Positions."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/StockPositionPageResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/stock-positions/{id}:
    get:
      operationId: getStockPosition
      summary: "Read one current Stock Position"
      security: [{ bearerAuth: [] }]
      x-required-permission: "STOCK:VIEW"
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: "Stock Position."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/StockPositionResponse' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/stock-positions/{id}/movements:
    get:
      operationId: listStockMovements
      summary: "Read append-only movement history"
      description: "Filters: movementType, businessDateFrom, businessDateTo. Sort allowlist is occurredAt, businessDate and id."
      security: [{ bearerAuth: [] }]
      x-required-permission: "STOCK:VIEW"
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - { name: movementType, in: query, schema: { type: string, enum: [PRODUCTION, DELIVERY, DELIVERY_REVERSAL, RETURN, DISPOSE] } }
        - { name: businessDateFrom, in: query, schema: { type: string, format: date } }
        - { name: businessDateTo, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          description: "Paged immutable movements."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/StockMovementPageResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/stock-positions/{id}/eligible-return-sources:
    get:
      operationId: listEligibleReturnSources
      summary: "Read only minimal delivered-line fields needed for a Return"
      description: "STOCK:RETURN is sufficient; this endpoint does not confer DELIVERY:VIEW."
      security: [{ bearerAuth: [] }]
      x-required-permission: "STOCK:RETURN"
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: "Sort allowlist is deliveryDate, deliveryNo and deliveryNoteItemId."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/EligibleReturnSourcePageResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/stock-positions/{id}/returns:
    post:
      operationId: recordStockReturn
      summary: "Append a customer Return against one eligible Delivery line"
      security: [{ bearerAuth: [] }]
      x-required-permission: "STOCK:RETURN"
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/StockReturnRequest' } } }
      responses:
        '200':
          description: "Movement appended and Stock Position reconciled."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/StockCommandResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/CommandConflict' }
        '410': { $ref: '#/components/responses/IdempotencyResultExpired' }
        '422': { $ref: '#/components/responses/UnprocessableCommand' }
  /api/v1/stock-positions/{id}/disposals:
    post:
      operationId: recordStockDisposal
      summary: "Append a finished-goods disposal"
      security: [{ bearerAuth: [] }]
      x-required-permission: "STOCK:DISPOSE"
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/StockDisposalRequest' } } }
      responses:
        '200':
          description: "Movement appended and Stock Position reconciled."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/StockCommandResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/CommandConflict' }
        '410': { $ref: '#/components/responses/IdempotencyResultExpired' }
        '422': { $ref: '#/components/responses/UnprocessableCommand' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "RS256 access JWT signed by an RSA-3072 key with a kid header and a 15 minute lifetime."
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: "Required for unsafe cookie-backed Auth calls. Reacquire it after login and logout."
    refreshCookie:
      type: apiKey
      in: cookie
      name: ERP_REFRESH
      description: "Rotating opaque refresh token. Secure HttpOnly SameSite=Strict and scoped to /api/v1/auth. Only its hash is stored server-side."
  headers:
    NoStore:
      description: "Always no-store for protected transaction and report responses."
      schema: { type: string, enum: [no-store] }
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: "Opaque client retry key, scoped by actor, method, and normalized path. The request hash includes body and If-Match. Retained for 24 hours."
      schema: { type: string, minLength: 1, maxLength: 120 }
    Page:
      name: page
      in: query
      description: "Zero-based page index."
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      description: "Page size."
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Sort:
      name: sort
      in: query
      description: "Repeatable allowlisted sort term written as field,direction. Repeat the parameter to add a secondary term such as a stable id tiebreaker."
      style: form
      explode: true
      schema:
        type: array
        items: { type: string, example: "code,asc" }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: "Quoted optimistic version of the addressed resource."
      schema: { type: string, pattern: '^"[0-9]+"$' }
  responses:
    BadRequest:
      description: "INVALID_REQUEST or VALIDATION_FAILED, including unknown request members."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Unauthenticated:
      description: "UNAUTHENTICATED, CSRF_INVALID or REFRESH_REAUTH_REQUIRED. The SPA may attempt one refresh for these codes."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Forbidden:
      description: "FORBIDDEN when effective permissions deny the operation, or PASSWORD_CHANGE_REQUIRED when a restricted challenge calls anything other than change-password or logout. Neither code may trigger a refresh attempt."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: "NOT_FOUND for the addressed resource or ownership path."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    CommandConflict:
      description: "VERSION_CONFLICT, IDEMPOTENCY_IN_PROGRESS, invalid transition, downstream activity, stock conflict, or another stable command conflict."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    IdempotencyResultExpired:
      description: "A quarantined legacy same-hash result cannot be replayed until its record is purged."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    UnprocessableCommand:
      description: "IDEMPOTENCY_KEY_REUSED or another semantically invalid command."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
  schemas:
    DecimalQuantity:
      type: string
      pattern: '^\d{1,14}(\.\d{1,4})?$'
      description: "Non-negative decimal quantity carried as a JSON string with up to four decimal places."
...
    ProblemDetail:
      type: object
      additionalProperties: true
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: "Stable machine-readable error code."
          enum:
            - INVALID_REQUEST
            - VALIDATION_FAILED
            - UNAUTHENTICATED
            - CSRF_INVALID
            - REFRESH_REAUTH_REQUIRED
            - PASSWORD_CHANGE_REQUIRED
            - FORBIDDEN
            - NOT_FOUND
            - VERSION_CONFLICT
            - DUPLICATE_BUSINESS_KEY
            - MASTER_IN_USE
            - RECOVERY_ADMIN_REQUIRED
            - EXCHANGE_RATE_MISSING
            - RATE_LIMITED
            - INTERNAL_ERROR
            - IF_MATCH_REQUIRED
            - INVALID_IF_MATCH
            - IDEMPOTENCY_KEY_REQUIRED
            - INVALID_IDEMPOTENCY_KEY
            - IDEMPOTENCY_KEY_REUSED
            - IDEMPOTENCY_IN_PROGRESS
            - IDEMPOTENCY_RESULT_EXPIRED
            - INVALID_STATE_TRANSITION
            - DOWNSTREAM_ACTIVITY_EXISTS
            - GROUP_MEMBERSHIP_INVALID
            - PRODUCTION_ALREADY_FINISHED
            - INSUFFICIENT_STOCK
            - DELIVERY_ALREADY_POSTED
            - DELIVERY_ALREADY_REVERSED
            - DELIVERY_HAS_RETURNS
            - RETURN_LIMIT_EXCEEDED
            - DISPOSAL_EXCEEDS_STOCK
            - REPORT_LIMIT_EXCEEDED
            - REPORT_BUSY
        traceId: { type: string }
        fieldErrors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    PageMetadata:
      type: object
      additionalProperties: false
      required: [number, size, totalElements, totalPages]
      properties:
        number: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    AppliedFilters:
      type: object
      additionalProperties: true
      description: "Echo of the accepted allowlisted filters only."
    AppliedSort:
      type: array
      items: { type: string }
      description: "Echo of the normalized allowlisted sort terms."
...
    StockPositionResponse:
      type: object
      additionalProperties: false
      required: [id, version, productionOrderId, buyerOrderItemId, customerId, currencyCode, uomCode, orderQty, producedQty, deliveredQty, returnedQty, disposedQty, currentQty, orderBalanceQty, updatedAt]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 0 }
        productionOrderId: { type: string, format: uuid }
        buyerOrderItemId: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        currencyCode: { type: string }
        uomCode: { type: string }
        orderQty: { $ref: '#/components/schemas/DecimalQuantity' }
        producedQty: { $ref: '#/components/schemas/DecimalQuantity' }
        deliveredQty: { $ref: '#/components/schemas/DecimalQuantity' }
        returnedQty: { $ref: '#/components/schemas/DecimalQuantity' }
        disposedQty: { $ref: '#/components/schemas/DecimalQuantity' }
        currentQty: { $ref: '#/components/schemas/DecimalQuantity' }
        orderBalanceQty: { $ref: '#/components/schemas/DecimalQuantity' }
        updatedAt: { type: string, format: date-time }
    StockMovementResponse:
      type: object
      additionalProperties: false
      required: [id, movementType, quantitySigned, balanceAfter, businessDate, sourceType, sourceId, occurredAt]
      properties:
        id: { type: string, format: uuid }
        movementType: { type: string, enum: [PRODUCTION, DELIVERY, DELIVERY_REVERSAL, RETURN, DISPOSE] }
        quantitySigned: { type: string, pattern: '^-?\d{1,14}(\.\d{1,4})?$' }
        balanceAfter: { $ref: '#/components/schemas/DecimalQuantity' }
        businessDate: { type: string, format: date }
        sourceType: { type: string }
        sourceId: { type: string, format: uuid }
        sourceItemId: { type: string, format: uuid }
        reason: { type: string }
        occurredAt: { type: string, format: date-time }
    EligibleReturnSourceResponse:
      type: object
      additionalProperties: false
      required: [deliveryNoteItemId, deliveryNo, deliveryDate, deliveredQty, netReturnableQty, uomCode]
      properties:
        deliveryNoteItemId: { type: string, format: uuid }
        deliveryNo: { type: string }
        deliveryDate: { type: string, format: date }
        deliveredQty: { $ref: '#/components/schemas/DecimalQuantity' }
        netReturnableQty: { $ref: '#/components/schemas/DecimalQuantity' }
        uomCode: { type: string }
    StockReturnRequest:
      type: object
      additionalProperties: false
      required: [deliveryNoteItemId, quantity, businessDate, reason]
      properties:
        deliveryNoteItemId: { type: string, format: uuid }
        quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        businessDate: { type: string, format: date }
        reason: { type: string, minLength: 1, maxLength: 200 }
    StockDisposalRequest:
      type: object
      additionalProperties: false
      required: [quantity, businessDate, reason]
      properties:
        quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        businessDate: { type: string, format: date }
        reason: { type: string, minLength: 1, maxLength: 200 }
    StockCommandResponse:
      type: object
      additionalProperties: false
      required: [stockPosition, movement]
      properties:
        stockPosition: { $ref: '#/components/schemas/StockPositionResponse' }
        movement: { $ref: '#/components/schemas/StockMovementResponse' }
...
    StockPositionPageResponse:
      type: object
      additionalProperties: false
      required: [items, page, filters, sort]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/StockPositionResponse' } }
        page: { $ref: '#/components/schemas/PageMetadata' }
        filters: { $ref: '#/components/schemas/AppliedFilters' }
        sort: { $ref: '#/components/schemas/AppliedSort' }
    StockMovementPageResponse:
      type: object
      additionalProperties: false
      required: [items, page, filters, sort]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/StockMovementResponse' } }
        page: { $ref: '#/components/schemas/PageMetadata' }
        filters: { $ref: '#/components/schemas/AppliedFilters' }
        sort: { $ref: '#/components/schemas/AppliedSort' }
    EligibleReturnSourcePageResponse:
      type: object
      additionalProperties: false
      required: [items, page, filters, sort]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/EligibleReturnSourceResponse' } }
        page: { $ref: '#/components/schemas/PageMetadata' }
        filters: { $ref: '#/components/schemas/AppliedFilters' }
        sort: { $ref: '#/components/schemas/AppliedSort' }
...